Added the swagger-ui api docs server into our build of the api server. Basically this was copying the dist folder out of their repos as they recommend here: https://github.com/swagger-api/swagger-ui#download
Also made the default page something better.

Updated the readme for how to build swagger.json (we need a better
name, but this is ok for now)

Added a pre-built swagger.json (following the instructions in the
tools/README.md file) to the git repository so the api-docs
server can work.