JSON style timestamps should now work right with news. We're using JODA time everywhere... If you ask why JODA time and not the new Java8 time?  Simple, cuz the new one doesn't work right with Jackson for serializing from JSON default date format...
Added ConverterUtis for use in the backend to convert to and from
whatever we use on model (org.joda.time.DateTime now, maybe
java.time.ZonedDateTime one day), so that you just have 1 place
to fix up vs. N number of callsites that need to be addressed.

The date format in Job is still just YYYY-MM-DD, but it's easy
to change to timestamps as needed.